backend/pyproject.toml
[project]
name = "bdc-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "snowflake-connector-python>=3.12.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
    "cryptography>=43.0.0",
]
